refactor: move guard functions for owc-json.ts to separate file #91
