+++ projects/services-ogc/src/lib/owc/types/owc-json.ts
@@ -8,16 +8,6 @@
  */
 import * as GeoJSON from 'geojson';
 
-
-function trueForAll(list: any[], predicate: (o: any) => boolean): boolean {
-  for (const entry of list) {
-    if (!predicate(entry)) {
-      return false;
-    }
-  }
-  return true;
-}
-
 /**
  * The OWS Context describes Metadata, API, Time Range
  * http://www.owscontext.org/owc_user_guide/C0_userGuide.html#truethe-ows-context-document-structure
@@ -68,20 +58,6 @@ export interface IOwsContext extends GeoJSON.FeatureCollection<GeoJSON.GeometryO
   [k: string]: any;
 }
 
-export function isIOwsContext(object: any): object is IOwsContext {
-   return 'id' in object
-     && 'properties' in object
-     && 'links' in object.properties // && trueForAll(object.properties.links, isIOwsLinks)
-     && 'lang' in object.properties
-     && 'title' in object.properties
-     && 'updated' in object.properties
-     && (object.properties.authors ? trueForAll(object.properties.authors, isIOwsAuthor) : true)
-     && (object.properties.creator ? isIOwsCreator(object.properties.creator) : true)
-     && (object.properties.display ? trueForAll(object.properties.display, isIOwsCreatorDisplay): true)
-     && (object.properties.categories ? trueForAll(object.properties.categories, isIOwsCategory) : true)
-     && 'features' in object && trueForAll(object.features, isIOwsResource);
- }
-
 /**
  * Each layer (a.k.a. feature) in a context document is known as a ‘Resource’
  * A Resource reference a set of geospatial information to be treated as a logical element.
@@ -103,12 +79,6 @@ export interface IOwsResource extends GeoJSON.Feature {
   properties: IOwsResourceProperties;
   [k: string]: any;
 }
-
-export function isIOwsResource(object: any): object is IOwsResource {
-  return 'id' in object
-    && 'properties' in object && isIOwsResourceProperties(object.properties);
-}
-
 export interface IOwsResourceProperties {
   /** Title given to the Context resource */
   title: string;
@@ -143,15 +113,6 @@ export interface IOwsResourceProperties {
   [k: string]: any;
 }
 
-export function isIOwsResourceProperties(object: any): object is IOwsResourceProperties {
-  return 'title' in object
-    && 'updated' in object
-    && (object.authors ? trueForAll(object.authors, isIOwsAuthor) : true)
-    && (object.offerings ? trueForAll(object.offerings, isIOwsOffering) : true)
-    && (object.categories ? trueForAll(object.categories, isIOwsCategory) : true);
-}
-
-
 /**
  * In reality a resource can be realized in a number of different ways, and so an OWC document allows various options to be specified.
  * These are known as offerings.
@@ -184,23 +145,10 @@ export interface IOwsOffering {
   [k: string]: any;
 }
 
-export function isIOwsOffering(object: any): object is IOwsOffering {
-  return 'code' in object
-    && (object.operations ? trueForAll(object.operations, isIOwsOperation) : true)
-    && (object.contents ? trueForAll(object.contents, isIOwsContent) : true)
-    && (object.styles ? trueForAll(object.styles, isIOwsStyleSet) : true)
-}
-
 export interface IOwsCreator {
   title?: string;
   uri?: string;
   version?: string;
-}
-
-export function isIOwsCreator(object: any): object is IOwsCreator {
-  return 'title' in object
-    || 'uri' in object
-    || 'version' in object;
 }
 
 export interface IOwsAuthor {
@@ -211,23 +159,11 @@ export interface IOwsAuthor {
   [k: string]: any;
 }
 
-export function isIOwsAuthor(object: any): object is IOwsAuthor {
-  return 'name' in object
-    || 'email' in object
-    || 'uri' in object;
-}
-
 export interface IOwsCategory {
   scheme?: string;
   /** Category related to this context document. It MAY have a related code-list that is identified by the scheme attribute */
   term?: string;
   label?: string;
-}
-
-export function isIOwsCategory(object: any): object is IOwsCategory {
-  return 'scheme' in object
-    || 'term' in object
-    || 'label' in object;
 }
 
 export interface IOwsLinks {
@@ -239,10 +175,6 @@ export interface IOwsLinks {
   alternates?: string;
   lang?: LangString;
   [k: string]: any;
-}
-
-export function isIOwsLinks(object: any): object is IOwsLinks {
-  return 'rel' in object;
 }
 
 export interface IOwsCreatorApplication {
@@ -261,12 +193,6 @@ export interface IOwsCreatorDisplay {
    */
   mmPerPixel?: number;
   [k: string]: any;
-}
-
-export function isIOwsCreatorDisplay(object: any): object is IOwsCreatorDisplay {
-  return 'pixelWidth' in object
-    || 'pixelHeight' in object
-    || 'mmPerPixel' in object;
 }
 
 /**
@@ -289,13 +215,6 @@ export interface IOwsOperation {
   [k: string]: any;
 }
 
-export function isIOwsOperation(object: any): object is IOwsOperation {
-  return 'code' in object
-    && 'method' in object
-    && (object.request ? isIOwsContent(object.request) : true)
-    && (object.result ? isIOwsContent(object.result) : true);
-}
-
 export interface IOwsContent {
   /** MIME type of the Content */
   type: string;
@@ -306,10 +225,6 @@ export interface IOwsContent {
   [k: string]: any;
 }
 
-export function isIOwsContent(object: any): object is IOwsContent {
-  return 'type' in object;
-}
-
 export interface IOwsStyleSet {
   name: string;
   title: string;
@@ -318,11 +233,6 @@ export interface IOwsStyleSet {
   legendURL?: string;
   content?: IOwsContent;
   [k: string]: any;
-}
-
-export function isIOwsStyleSet(object: any): object is IOwsStyleSet {
-  return 'name' in object
-    && 'title' in object;
 }
 
 /** ISO-8601 format e.g. YYYY-MM-DDThh:mm:ssZ or YYYY-MM-DDThh:mm:ssZ/YYYY-MM-DDThh:mm:ssZ */
