+++ projects/services-ogc/src/public-api.ts
@@ -5,6 +5,7 @@
 export * from './lib/services-ogc.module';
 export * from './lib/owc/owc-json.service';
 export * from './lib/owc/types/owc-json';
+export * from './lib/owc/types/owc-json.utils';
 export * from './lib/owc/types/eoc-owc-json';
 export * from './lib/wps/wpsclient';
 export * from './lib/wmts/wmtsclient.service';
